+++ examples/mass_example.py
@@ -1,9 +1,13 @@
 from peptacular.mass import calculate_mass, calculate_mz
 
-# Calculate mass (using monoisotopic mass)
+# Calculate mass (monoisotopic)
 peptide_mass = calculate_mass('PEP(1.0)TIDE[2.0]', charge=2)
 assert peptide_mass == 804.3745169602499
 
-# Calculate m/z (using average mass)
+# Calculate m/z (monoisotopic)
+peptide_mz = calculate_mz('PEP(1.0)TIDE[2.0]', charge=2)
+assert peptide_mz == 402.18725848012497
+
+# Calculate m/z (average)
 peptide_mz = calculate_mz('PEP(1.0)TIDE[2.0]', charge=2, monoisotopic=False)
 assert peptide_mz == 402.42130880862