+++ src/peptacular/constants.py
@@ -4,6 +4,8 @@ AMINO_ACIDS = 'ACDEFGHIKLMNPQRSTVWY'
 
 PROTON_MASS = 1.00727646677
 
+AMINO_ACIDS = {'S', 'A', 'O', 'Y', 'R', 'P', 'N', 'H', 'E', 'J', 'G', 'F', 'V', 'D', 'I', 'K', 'L', 'C', 'T', 'Q', 'U', 'M', 'W'}
+
 MONO_ISOTOPIC_ATOMIC_MASSES = {
     "HYDROGEN": 1.00782503207,
     "PROTON": PROTON_MASS,